replace OutputStream in downloader with ProgressOutputStream to track download progress
