+++ src/main/java/com/dropbox/core/DbxDownloader.java
@@ -1,6 +1,7 @@
 package com.dropbox.core;
 
 import com.dropbox.core.util.IOUtil;
+import com.dropbox.core.util.ProgressOutputStream;
 
 import java.io.Closeable;
 import java.io.InputStream;
@@ -39,6 +40,7 @@ import java.io.OutputStream;
 public class DbxDownloader<R> implements Closeable {
     private final R result;
     private final InputStream body;
+    private IOUtil.ProgressListener progressListener;
 
     private boolean closed;
 
@@ -118,6 +120,22 @@ public class DbxDownloader<R> implements Closeable {
         }
 
         return result;
+    }
+
+    /**
+     * This method is the same as {@link #download(OutputStream)} except for allowing to track
+     * download progress.
+     *
+     * @param out {@code OutputStream} to write response body to
+     * @param progressListener {@code IOUtil.ProgressListener} to track the download progress.
+     * @return Response from server
+     * @throws DbxException if an error occurs reading the response or response body.
+     * @throws IOException if an error occurs writing the response body to the output stream.
+     */
+    public R download(OutputStream out, IOUtil.ProgressListener progressListener)
+            throws DbxException,  IOException {
+        this.progressListener = progressListener;
+        return download(new ProgressOutputStream(out, progressListener));
     }
 
     /**
